# Conflicts:
#	Cargo.toml

+++ kernels/qwen36_moe_persistent/batched_prefill_attn_full.cuh
@@ -54,7 +54,7 @@ __device__ __forceinline__ float to_float(T x);
 
 template <>
 __device__ __forceinline__ float to_float<hip_bfloat16>(hip_bfloat16 x) {
-    return __bfloat162float(x);
+    return static_cast<float>(x);
 }
 
 // Grid:  (ceil(q_len / BM), q_heads, batch)
